Allow default view_content_link on Mailer class
This is something that @jamesmichiemo noticed we needed since we
was setting this in mailer methods all over the place.

You can now set a default view_content_link in either your Mailer class or
your BaseMailer class. Just like the other defaults, this one cascades
down and can be overidden at the Mailer class level or the method level.