Instatiate the mailer class when sending later.
Out of the box this code is essentially a no-op but if the subclass
has overridden functionality in the base mailer this ensures that
overwritten functionality happens.

I use this primarily to override `deliver_now` so that I can extract
the mandril response the the REST call and store the message id's
locally. Example:

    # Override to record the message id with each lead so we know when we last
    # communicated with a given lead. This also works even if `deliver_later` is
    # called as `deliver_later` internally calls `deliver_now`.
    def deliver_now
      super.tap do |results|
        emails = results.collect {|r| r['email']}
        leads = Lead.where(email: emails).index_by &:email
        for result in results
          leads[result['email']].communications.create! message_id: result['_id']
        end
      end
    end