Keep API compatible.
In case someone is using the stock job's directly move the mailer to
be the last argument and make it optional. If not provided then the
old behavior will take place. If provided then the given mailer will
be used.

When this library creates the job then the argument is provided ensuring
the right class is instantiated.