Enfoce use of inline adapter.
The specs assume the inline adapter is being used. This was true prior
to Rails 5 but the new default in Rails 5 is the async adapter.

We could possibly update the failing tests to be compatible with the
async adapter but I think the inline adapter is better suited for
tests generally anyway as it makes the tests easier and the difference
in parity is not generally that worriesome considering most people don't
use either adapter for production (so neither adapter has parity).