Reduce redefinations in offline support for testing.
The `deliver` method does not need to be redefined as it is the same
as the real version (they both proxy to `deliver_now`).

The `deliver_later` does not need to be overwritten as after it goes
through ActiveJob it will end up calling the redefined `deliver_now`.

This does result in a slight change in behavior. If the ActiveJob
adapter is not `inline` then the `MandrillMailer.deliveries` will not
be yet populated. I see this as a good thing as it more closely emulates
the original behavior (i.e. nothing happens now). This is also in parity
with ActionMailer (the `deliverables` array is not populated if using
`deliver_later` and not using the `inline` adapter).

In general I think the `inline` adapter should be used for testing for
this reason (simplicity in testing) but having it go through the real
ActiveJob allows the calling code to control this. Also it will then
kick off the instrumation events (`enqueue.active_job`).